fixed runtime bugs

+++ src/runtime_src/xdp/profile/database/static_info_database.h
@@ -1066,8 +1066,8 @@ class aie_cfg_tile
     }
 
     inline void addAIECounter(uint64_t deviceId, uint32_t i, uint16_t col,
-            uint16_t r, uint8_t num, uint8_t start,
-            uint8_t end, uint8_t reset, double freq,
+            uint16_t r, uint8_t num, uint16_t start,
+            uint16_t end, uint8_t reset, double freq,
             const std::string& mod,
             const std::string& aieName)
     {
